backend/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
requests==2.31.0
pytesseract==0.3.10
Pillow==12.1.1
pdf2image==1.16.3
python-dotenv==1.0.0
Werkzeug==2.2.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
bcrypt==4.0.1